embedded/requirements.txt
astroid==2.3.0
atomicwrites==1.3.0
attrs==19.2.0
bottle==0.12.19
Cerberus==1.3.2
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.1
contextlib2==0.5.5
cycler==0.10.0
docker==4.1.0
elasticsearch==7.1.0
entrypoints==0.3
flake8==3.7.9
flasgger==0.9.3
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==2.6.0
keyring==19.2.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
marshmallow==2.20.5
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==7.2.0
mplcursors==0.3
numpy==1.17.2
packaging==19.2
platformio==5.0.4
pluggy==0.13.0
py==1.10.0
pycodestyle==2.5.0
pyelftools==0.25
pyflakes==2.1.1
pylint==2.4.1
pyparsing==2.4.2
pyserial==3.4
pytest==5.2.0
python-dateutil==2.8.0
PyYAML==5.4
requests==2.22.0
schema==0.7.1
semantic-version==2.8.2
six==1.12.0
tabulate==0.8.5
tinydb==3.15.1
typed-ast==1.4.0
urllib3==1.26.5
virtualenv==16.6.0
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.16.0
wheel==0.34.2
wrapt==1.11.2
zipp==0.6.0
